+++ module/actes-generique/action/SAEEnvoiActes.class.php
@@ -40,7 +40,7 @@ class SAEEnvoiActes extends ActionExecutor
         $sae = $this->getConnecteur('SAE');
 
         $fluxData = new FluxDataSedaActes($donneesFormulaire);
-        $bordereau = $actesSEDA->getBordereauNG($fluxData);
+        $bordereau = $actesSEDA->getBordereau($fluxData);
         $donneesFormulaire->addFileFromData('sae_bordereau', "bordereau.xml", $bordereau);
         $transferId = $sae->getTransferId($bordereau);
         $donneesFormulaire->setData("sae_transfert_id", $transferId);